Refactoring CIS scoring & structure logic
